% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_edges}
\alias{construct_edges}
\title{Get all possible edges for a graph with p vertices}
\usage{
construct_edges(p)
}
\arguments{
\item{p}{Number of vertices}
}
\description{
Get all possible edges for a graph with p vertices
}
\examples{
construct_edges(5) # 10 edges
}
