% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giigTests-package.R
\docType{package}
\name{giigTests-package}
\alias{giigTests}
\alias{giigTests-package}
\title{giigTests: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Przemyslaw Chojecki \email{premysl.choj@gmail.com}

}
\keyword{internal}
