% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_in_list}
\alias{is_in_list}
\title{Is \code{el} the element of the \code{list}.}
\usage{
is_in_list(el, list)
}
\description{
Is \code{el} the element of the \code{list}.
}
\examples{
is_in_list(c(1,2,3), list(c(1,2,3), c(3,2,1))) # TRUE
is_in_list(c(1,2,3), list(c(2,3,1), c(3,2,1))) # FALSE
}
