% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfect_elimination_orderings.R
\name{is_simplicial}
\alias{is_simplicial}
\title{Find out if the vertev v is simplicial}
\usage{
is_simplicial(v, adj)
}
\arguments{
\item{v}{vertex of interest}

\item{adj}{Adjacency matrix of a graph of interest}
}
\description{
Find out if the vertev v is simplicial
}
\examples{
M <- matrix(c(
  0, 0, 0, 0,
  0, 0, 1, 0,
  0, 1, 0, 1,
  0, 0, 1, 0
), nrow = 4) # 1 2-3-4
sapply(1:4, function(v) {
  is_simplicial(v, M)
}) # TRUE, TRUE, FALSE, TRUE
}
