% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meet_and_join.R
\name{join}
\alias{join}
\title{Make join of two graphs according to Gehrmann}
\usage{
join(vPartition1, ePartition1, vPartition2, ePartition2)
}
\description{
Make join of two graphs according to Gehrmann
}
\examples{
vPartition1 <- list(c(1, 3), c(2, 4))
vPartition2 <- list(c(1, 3), c(2), c(4))
ePartition1 <- list(c(1, 6), c(3, 4))
ePartition2 <- list(c(1, 4), c(3, 6), c(2))

join(vPartition1, ePartition1, vPartition2, ePartition2)
}
