% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_factor_graph.R
\name{make_factor_graph_for_nauty}
\alias{make_factor_graph_for_nauty}
\title{Change calculate the factor graph and make the graph for nauty}
\usage{
make_factor_graph_for_nauty(vPartition, ePartition)
}
\description{
nauty numbers vertices from 0, not from 1.
}
\examples{
vPartition <- list(c(1, 2), c(3, 4))
ePartition <- list(1, c(3, 4), 6)

make_factor_graph_for_nauty(vPartition, ePartition)
}
