% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PartitionToEdgeColorClasses}
\alias{PartitionToEdgeColorClasses}
\title{Change the partition from indexes of the edges to edges itself.}
\usage{
PartitionToEdgeColorClasses(partition, edges)
}
\description{
Change the partition from indexes of the edges to edges itself.
}
\examples{
p <- 4
edges <- construct_edges(p)
allPartitions <- partitions::listParts(p*(p-1)/2) # All possible colorings of edges
myPartition <- allPartitions[[7]] # this is the partition of indices of the edges
ECC <- PartitionToEdgeColorClasses(myPartition, edges) # this is the partition of the edges
}
