% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfect_elimination_orderings.R
\name{perfect_elimination_orderings}
\alias{perfect_elimination_orderings}
\title{Find all the Perfect Elimination Orderings (PEOs)}
\usage{
perfect_elimination_orderings(adj)
}
\arguments{
\item{adj}{Adjacency matrix of a graph of interest}
}
\description{
Find all the Perfect Elimination Orderings (PEOs)
}
\examples{
M <- matrix(c(
  0, 0, 0, 0,
  0, 0, 1, 0,
  0, 1, 0, 1,
  0, 0, 1, 0
), nrow = 4) # 1 2-3-4
perfect_elimination_orderings(M) # list of 16 PEOs
}
